Module alloy_node_bindings::geth

Utilities for launching a go-ethereum dev-mode instance.

Structs

DevOptions

Configuration options that can be set in dev mode.

Geth

Builder for launching geth.

GethInstance

A geth instance. Will close the instance when dropped.

PrivateNetOptions

Configuration options that cannot be set in dev mode.

Enums

GethError

Errors that can occur when working with the Geth.

GethInstanceError

Errors that can occur when working with the GethInstance.

GethMode

Whether or not geth is in dev mode and configuration options that depend on the mode.

Constants

API 🔒

The exposed APIs

GETH 🔒

The geth command

GETH_DIAL_LOOP_TIMEOUT 🔒

Timeout for waiting for geth to add a peer.

GETH_STARTUP_TIMEOUT 🔒

How long we will wait for geth to indicate that it is ready.

Functions

extract_endpoint 🔒

extract_value 🔒