Crate alloy_node_bindings

alloy-node-bindings

Ethereum execution-layer client bindings.

Re-exports

pub use anvil::Anvil;

pub use anvil::AnvilInstance;

pub use geth::Geth;

pub use geth::GethInstance;

Modules

anvil

Utilities for launching an Anvil instance.

geth

Utilities for launching a go-ethereum dev-mode instance.

Constants

EIP1559_FEE_ESTIMATION_DEFAULT_PRIORITY_FEE

The default max priority fee per gas, used in case the base fee is within a threshold.

EIP1559_FEE_ESTIMATION_PAST_BLOCKS

The number of blocks from the past for which the fee rewards are fetched for fee estimation.

EIP1559_FEE_ESTIMATION_PRIORITY_FEE_TRIGGER

The threshold for base fee below which we use the default priority fee, and beyond which we estimate an appropriate value for priority fee.

EIP1559_FEE_ESTIMATION_REWARD_PERCENTILE

The default percentile of gas premiums that are fetched for fee estimation.

EIP1559_FEE_ESTIMATION_THRESHOLD_MAX_CHANGE

The threshold max change/difference (in %) at which we will ignore the fee history values under it.

WEI_IN_ETHER

1 Ether = 1e18 Wei == 0x0de0b6b3a7640000 Wei

Functions

unused_port 🔒

A bit of hack to find an unused TCP port.